+++ android/app/src/main/java/com/onair/hearit/presentation/detail/adapter/PlayerDetailScriptViewHolder.kt
@@ -1,4 +1,4 @@
-package com.onair.hearit.presentation.detail
+package com.onair.hearit.presentation.detail.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
